+++ src/pharmpy/tools/__init__.py
@@ -9,6 +9,7 @@ __all__ = (
     'print_fit_summary',  # pyright: ignore [reportUnsupportedDunderAll]
     'rank_models',  # pyright: ignore [reportUnsupportedDunderAll]
     'read_results',  # pyright: ignore [reportUnsupportedDunderAll]
+    'read_modelfit_results',  # pyright: ignore [reportUnsupportedDunderAll]
     'retrieve_final_model',  # pyright: ignore [reportUnsupportedDunderAll]
     'retrieve_models',  # pyright: ignore [reportUnsupportedDunderAll]
     'run_allometry',  # pyright: ignore [reportUnsupportedDunderAll]
@@ -35,6 +36,7 @@ _run_keys = {
     'fit',
     'print_fit_summary',
     'rank_models',
+    'read_modelfit_results',
     'read_results',
     'retrieve_final_model',
     'retrieve_models',